app/client/requirements.txt
numpy
pillow
bentoml
fastapi[standard]
py-grpc-prometheus